backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
firebase-admin==6.5.0
torch>=2.6.0
transformers>=4.34.0
python-multipart>=0.0.6
geopy==2.4.1
httpx>=0.25.0
requests>=2.25.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.2 